+++ src/index.ts
@@ -1,8 +1,6 @@
 /**
  * index module
  */
-import * as redis from 'redis';
-
 import * as AggregateService from './service/aggregate';
 import * as NotificationService from './service/notification';
 import * as PerformanceService from './service/performance';
@@ -12,11 +10,6 @@ import * as TaskService from './service/task';
 import * as factory from '@tokyotower/factory';
 
 import * as repository from './repository';
-
-/**
- * Redis Cacheクライアント
- */
-export import redis = redis;
 
 export namespace service {
     export import notification = NotificationService;