+++ src/repository.ts
@@ -3,7 +3,6 @@
  * リポジトリ
  */
 import { MongoRepository as AggregateSaleRepo } from './repo/aggregateSale';
-import { RedisRepository as EventWithAggregationRepo } from './repo/event';
 import { MongoRepository as PerformanceRepo } from './repo/performance';
 import { MongoRepository as ReservationRepo } from './repo/reservation';
 import { MongoRepository as TaskRepo } from './repo/task';
@@ -12,11 +11,6 @@ import { MongoRepository as TaskRepo } from './repo/task';
  * 売上集計リポジトリ
  */
 export class AggregateSale extends AggregateSaleRepo { }
-
-/**
- * 集計データ付きイベントリポジトリ
- */
-export class EventWithAggregation extends EventWithAggregationRepo { }
 
 /**
  * パフォーマンスリポジトリ