+++ src/service/aggregate.ts
@@ -15,15 +15,20 @@ import { credentials } from '../credentials';
 
 const debug = createDebug('ttts-domain:service');
 
-const project: factory.project.IProject = {
+const project: cinerinoapi.factory.project.IProject = {
     typeOf: cinerinoapi.factory.organizationType.Project,
     id: <string>process.env.PROJECT_ID
 };
 
-const EVENT_AGGREGATION_EXPIRES_IN_SECONDS = (process.env.EVENT_AGGREGATION_EXPIRES_IN_SECONDS !== undefined)
-    ? Number(process.env.EVENT_AGGREGATION_EXPIRES_IN_SECONDS)
-    // tslint:disable-next-line:no-magic-numbers
-    : 86400;
+export enum SeatingType {
+    Normal = 'Normal',
+    Wheelchair = 'Wheelchair'
+}
+
+export enum TicketTypeCategory {
+    Normal = 'Normal',
+    Wheelchair = 'Wheelchair'
+}
 
 const WHEEL_CHAIR_NUM_ADDITIONAL_STOCKS = (process.env.WHEEL_CHAIR_NUM_ADDITIONAL_STOCKS !== undefined)
     ? Number(process.env.WHEEL_CHAIR_NUM_ADDITIONAL_STOCKS)
@@ -49,7 +54,6 @@ export function aggregateEventReservations(params: {
     id: string;
 }) {
     return async (repos: {
-        eventWithAggregation: repository.EventWithAggregation;
         performance: repository.Performance;
         reservation: repository.Reservation;
     }) => {
@@ -99,36 +103,6 @@ export function aggregateEventReservations(params: {
             await aggregateByEvent({ checkGates: checkGates, event: aggregatingEvent })(repos);
         }
         debug('aggregated', aggregatingEvents.map((e) => e.id));
-
-        // 不要な集計データをクリーンアップ
-        try {
-            await makeAggregationsExpired()(repos);
-        } catch (error) {
-            // no op
-        }
-    };
-}
-
-function makeAggregationsExpired() {
-    return async (repos: {
-        eventWithAggregation: repository.EventWithAggregation;
-        performance: repository.Performance;
-    }) => {
-        // 過去のイベントを検索
-        const startThrough = moment()
-            .add(-1, 'week')
-            .toDate();
-        const startFrom = moment(startThrough)
-            .add(-1, 'week')
-            .toDate();
-        const eventIds = await repos.performance.distinct('_id', {
-            startFrom: startFrom,
-            startThrough: startThrough
-        });
-
-        if (eventIds.length > 0) {
-            await repos.eventWithAggregation.deleteByIds({ ids: eventIds });
-        }
     };
 }
 
@@ -137,12 +111,12 @@ function makeAggregationsExpired() {
  */
 function aggregateByEvent(params: {
     checkGates: factory.place.checkinGate.IPlace[];
-    event: factory.performance.IPerformanceWithDetails;
+    event: factory.performance.IPerformance;
 }) {
     // tslint:disable-next-line:max-func-body-length
     return async (repos: {
-        eventWithAggregation: repository.EventWithAggregation;
         reservation: repository.Reservation;
+        performance: repository.Performance;
     }) => {
         const checkGates = params.checkGates;
         const performance = params.event;
@@ -153,9 +127,6 @@ function aggregateByEvent(params: {
                 typeOf: factory.chevre.reservationType.EventReservation,
                 reservationStatuses: [factory.chevre.reservationStatusType.ReservationConfirmed],
                 reservationFor: { id: performance.id }
-                // additionalProperty: {
-                //     $nin: [{ name: 'extra', value: '1' }]
-                // }
             },
             // 集計作業はデータ量次第で時間コストを気にする必要があるので、必要なフィールドのみ取得
             {
@@ -184,17 +155,17 @@ function aggregateByEvent(params: {
 
             // オファーごとの集計
             const offersAggregation = await Promise.all(offers.map(async (offer) => {
-                let ticketTypeCategory = factory.ticketTypeCategory.Normal;
+                let ticketTypeCategory = TicketTypeCategory.Normal;
                 if (Array.isArray(offer.additionalProperty)) {
                     const categoryProperty = offer.additionalProperty.find((p) => p.name === 'category');
                     if (categoryProperty !== undefined) {
-                        ticketTypeCategory = <factory.ticketTypeCategory>categoryProperty.value;
+                        ticketTypeCategory = <TicketTypeCategory>categoryProperty.value;
                     }
                 }
 
                 return {
                     id: <string>offer.id,
-                    remainingAttendeeCapacity: (ticketTypeCategory === factory.ticketTypeCategory.Wheelchair)
+                    remainingAttendeeCapacity: (ticketTypeCategory === TicketTypeCategory.Wheelchair)
                         ? remainingAttendeeCapacityForWheelchair
                         : remainingAttendeeCapacity,
                     reservationCount: reservations.filter((r) => r.reservedTicket.ticketType.id === offer.id).length
@@ -241,8 +212,8 @@ function aggregateByEvent(params: {
             };
             debug('aggregated!', aggregation);
 
-            // 保管
-            await repos.eventWithAggregation.store([aggregation], EVENT_AGGREGATION_EXPIRES_IN_SECONDS);
+            // パフォーマンスリポジトリにも保管
+            await saveAggregation2performance(aggregation)(repos);
         } catch (error) {
             // tslint:disable-next-line:no-console
             console.error('couldn\'t create aggregation on event', performance.id, error);
@@ -251,11 +222,45 @@ function aggregateByEvent(params: {
 }
 
 /**
+ * パフォーマンスコレクションに集計データを保管する
+ */
+function saveAggregation2performance(params: factory.performance.IPerformanceWithAggregation) {
+    return async (repos: {
+        performance: repository.Performance;
+    }) => {
+        // 値がundefinedの場合に更新しないように注意
+        const update: any = {
+            $set: {
+                updated_at: new Date(), // $setオブジェクトが空だとMongoエラーになるので
+                evServiceStatus: params.evServiceStatus,
+                onlineSalesStatus: params.onlineSalesStatus,
+                maximumAttendeeCapacity: params.maximumAttendeeCapacity,
+                remainingAttendeeCapacity: params.remainingAttendeeCapacity,
+                remainingAttendeeCapacityForWheelchair: params.remainingAttendeeCapacityForWheelchair,
+                reservationCount: params.reservationCount,
+                checkinCount: params.checkinCount,
+                reservationCountsByTicketType: params.reservationCountsByTicketType,
+                checkinCountsByWhere: params.checkinCountsByWhere,
+                tourNumber: (<any>params).tourNumber,
+                ...(Array.isArray(params.offers)) ? { offers: params.offers } : undefined
+            },
+            $unset: {
+                noExistingAttributeName: 1, // $unsetは空だとエラーになるので
+                ...(!Array.isArray(params.offers)) ? { offers: '' } : undefined
+            }
+        };
+
+        // 保管
+        await repos.performance.updateOne({ _id: params.id }, update);
+    };
+}
+
+/**
  * 残席数を集計する
  */
 function aggregateRemainingAttendeeCapacity(params: {
-    performance: factory.performance.IPerformanceWithDetails;
-    project: factory.project.IProject;
+    performance: factory.performance.IPerformance;
+    project: cinerinoapi.factory.project.IProject;
 }) {
     // tslint:disable-next-line:max-func-body-length
     return async (__: {
@@ -291,9 +296,9 @@ function aggregateRemainingAttendeeCapacity(params: {
         // maximumAttendeeCapacityは一般座席数
         const maximumAttendeeCapacity = sectionOffer.containsPlace.filter(
             (p) => {
-                return (typeof p.seatingType === 'string' && p.seatingType === factory.place.movieTheater.SeatingType.Normal)
+                return (typeof p.seatingType === 'string' && p.seatingType === SeatingType.Normal)
                     || (Array.isArray(p.seatingType) &&
-                        (<any>p.seatingType).includes(<string>factory.place.movieTheater.SeatingType.Normal));
+                        (<any>p.seatingType).includes(<string>SeatingType.Normal));
             }
         ).length;
         let remainingAttendeeCapacity = maximumAttendeeCapacity;
@@ -310,15 +315,15 @@ function aggregateRemainingAttendeeCapacity(params: {
 
             // 一般座席
             const normalSeats = availableSeats.filter(
-                (s) => (typeof s.seatingType === 'string' && s.seatingType === factory.place.movieTheater.SeatingType.Normal)
+                (s) => (typeof s.seatingType === 'string' && s.seatingType === SeatingType.Normal)
                     || (Array.isArray(s.seatingType) &&
-                        (<any>s.seatingType).includes(<string>factory.place.movieTheater.SeatingType.Normal))
+                        (<any>s.seatingType).includes(<string>SeatingType.Normal))
             );
             // 全車椅子座席
             const wheelChairSeats = availableSeats.filter(
-                (s) => (typeof s.seatingType === 'string' && s.seatingType === factory.place.movieTheater.SeatingType.Wheelchair)
+                (s) => (typeof s.seatingType === 'string' && s.seatingType === SeatingType.Wheelchair)
                     || (Array.isArray(s.seatingType) &&
-                        (<any>s.seatingType).includes(<string>factory.place.movieTheater.SeatingType.Wheelchair))
+                        (<any>s.seatingType).includes(<string>SeatingType.Wheelchair))
             );
 
             const seats = sectionOffer.containsPlace;
@@ -332,7 +337,7 @@ function aggregateRemainingAttendeeCapacity(params: {
                     seatNumber: s.branchCode
                 };
             });
-            // const unavailableSeats = await repos.stock.findUnavailableOffersByEventId({ eventId: params.performance.id });
+
             const unavailableSeatNumbers = unavailableSeats.map((s) => s.seatNumber);
             debug('unavailableSeatNumbers:', unavailableSeatNumbers.length);
 
@@ -358,7 +363,7 @@ function aggregateRemainingAttendeeCapacity(params: {
                     }
                 }
 
-                return ticketTypeCategory === factory.ticketTypeCategory.Wheelchair;
+                return ticketTypeCategory === TicketTypeCategory.Wheelchair;
             });
             if (wheelChairOffer !== undefined && wheelChairOffer.availability === factory.chevre.itemAvailability.OutOfStock) {
                 remainingAttendeeCapacityForWheelchair = 0;
@@ -422,11 +427,11 @@ function aggregateCheckinCount(
         return {
             where: checkinGate.identifier,
             checkinCountsByTicketType: offers.map((offer) => {
-                let ticketTypeCategory = factory.ticketTypeCategory.Normal;
+                let ticketTypeCategory = TicketTypeCategory.Normal;
                 if (Array.isArray(offer.additionalProperty)) {
                     const categoryProperty = offer.additionalProperty.find((p) => p.name === 'category');
                     if (categoryProperty !== undefined) {
-                        ticketTypeCategory = <factory.ticketTypeCategory>categoryProperty.value;
+                        ticketTypeCategory = <TicketTypeCategory>categoryProperty.value;
                     }
                 }