+++ src/service/performance.ts
@@ -3,7 +3,6 @@ import * as factory from '@tokyotower/factory';
 import * as createDebug from 'debug';
 import * as moment from 'moment-timezone';
 
-import { RedisRepository as EventWithAggregationRepo } from '../repo/event';
 import { MongoRepository as PerformanceRepo } from '../repo/performance';
 import { MongoRepository as TaskRepo } from '../repo/task';
 
@@ -31,8 +30,7 @@ export interface ISearchResult {
 }
 
 export type ISearchOperation<T> = (
-    performanceRepo: PerformanceRepo,
-    eventWithAggregationRepo: EventWithAggregationRepo
+    performanceRepo: PerformanceRepo
 ) => Promise<T>;
 
 // 作成情報取得
@@ -156,8 +154,7 @@ export function importFromCinerino(params: factory.chevre.event.IEvent<factory.c
 export function search(searchConditions: factory.performance.ISearchConditions): ISearchOperation<ISearchResult> {
     // tslint:disable-next-line:max-func-body-length
     return async (
-        performanceRepo: PerformanceRepo,
-        eventWithAggregationRepo: EventWithAggregationRepo
+        performanceRepo: PerformanceRepo
     ) => {
         // 作品件数取得
         const filmIds = await performanceRepo.distinct('superEvent.workPerformed.identifier', searchConditions);
@@ -176,13 +173,8 @@ export function search(searchConditions: factory.performance.ISearchConditions):
         });
         debug(performances.length, 'performances found.');
 
-        // 空席情報を追加
-        const eventsWithAggregation = await eventWithAggregationRepo.findAll();
-        debug(eventsWithAggregation.length, 'eventsWithAggregation found.');
-
         const data: factory.performance.IPerformanceWithAvailability[] = performances.map((performance) => {
             const ticketTypes = (performance.ticket_type_group !== undefined) ? performance.ticket_type_group.ticket_types : [];
-            const eventWithAggregation = eventsWithAggregation.find((e) => e.id === performance.id);
 
             let tourNumber: string = (<any>performance).tour_number; // 古いデーターに対する互換性対応
             if (performance.additionalProperty !== undefined) {
@@ -197,18 +189,16 @@ export function search(searchConditions: factory.performance.ISearchConditions):
                 open_time: moment(performance.doorTime).tz('Asia/Tokyo').format('HHmm'),
                 start_time: moment(performance.startDate).tz('Asia/Tokyo').format('HHmm'),
                 end_time: moment(performance.endDate).tz('Asia/Tokyo').format('HHmm'),
-                // start_date: performance.startDate,
-                // end_date: performance.endDate,
-                seat_status: (eventWithAggregation !== undefined)
-                    ? eventWithAggregation.remainingAttendeeCapacity
+                seat_status: (typeof performance.remainingAttendeeCapacity === 'number')
+                    ? performance.remainingAttendeeCapacity
                     : undefined,
                 tour_number: tourNumber,
-                wheelchair_available: (eventWithAggregation !== undefined)
-                    ? eventWithAggregation.remainingAttendeeCapacityForWheelchair
+                wheelchair_available: (typeof performance.remainingAttendeeCapacityForWheelchair === 'number')
+                    ? performance.remainingAttendeeCapacityForWheelchair
                     : undefined,
                 ticket_types: ticketTypes.map((ticketType) => {
-                    const offerAggregation = (eventWithAggregation !== undefined && eventWithAggregation.offers !== undefined)
-                        ? eventWithAggregation.offers.find((o) => o.id === ticketType.id)
+                    const offerAggregation = (Array.isArray(performance.offers))
+                        ? performance.offers.find((o) => o.id === ticketType.id)
                         : undefined;
 
                     const unitPriceSpec = ticketType.priceSpecification;
@@ -232,6 +222,7 @@ export function search(searchConditions: factory.performance.ISearchConditions):
             };
 
             return {
+                ...performance,
                 id: performance.id,
                 doorTime: performance.doorTime,
                 startDate: performance.startDate,
@@ -245,11 +236,11 @@ export function search(searchConditions: factory.performance.ISearchConditions):
                     ? performance.ttts_extension.online_sales_status
                     : factory.performance.OnlineSalesStatus.Normal,
                 maximumAttendeeCapacity: MAXIMUM_ATTENDEE_CAPACITY,
-                remainingAttendeeCapacity: (eventWithAggregation !== undefined)
-                    ? eventWithAggregation.remainingAttendeeCapacity
+                remainingAttendeeCapacity: (typeof performance.remainingAttendeeCapacity === 'number')
+                    ? performance.remainingAttendeeCapacity
                     : undefined,
-                remainingAttendeeCapacityForWheelchair: (eventWithAggregation !== undefined)
-                    ? eventWithAggregation.remainingAttendeeCapacityForWheelchair
+                remainingAttendeeCapacityForWheelchair: (typeof performance.remainingAttendeeCapacityForWheelchair === 'number')
+                    ? performance.remainingAttendeeCapacityForWheelchair
                     : undefined,
                 extension: (performance.ttts_extension !== undefined) ? performance.ttts_extension : <any>{},
                 additionalProperty: performance.additionalProperty,