+++ CHANGELOG.md
@@ -10,9 +10,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- イベント集計データをperformancesコレクションにも保管するように調整
+- update @tokyotower/factory
+
 ### Deprecated
 
 ### Removed
+
+- 一時イベントリポジトリを削除
+- redisクライアントのエクスポートを削除
 
 ### Fixed